Moved deterministic computation inside compute_rtgm
